requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.23.4
pandas==1.5.0
psycopg2==2.9.4
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0
